Fixed out-of-sequence file name substitution, added log paths.
The script was replacing part of the .bam file path before calling the
JAR, which resulted in a file not found error.  The substitution was
moved to directly before the print to the resultant CSV file whose
content necessitates the substitution in the first place.

The log paths were also made absolute (although these should eventually
be put in the configuration file - hard-coded paths should be avoided in
source code altogether).