Uses metadata to determine parameters; changed wrapper method logic.
TUI now makes use of the metadata.txt file to determine which parameters
are required for each class called and the associated validating regex
for each parameter.

The wrapper logic method is much simpler, since the knowledge of the
parameters is decoupled from this module as a result of the metadata
file.

The method to prompt for input and to set instance members has been
changed to reflect the new data structure that stores the input from the
metadata file.