Initial commmit.
Originally combined with raw birdseed to GELI conversion.  With the
implementation of the Scheduler class, these needed to be removed to
different classes because parallel job submission would mean that the
raw birdseed to GELI conversion would not have time to complete before
the BS to GELI conversion is called.