Implements Samples data structure, decouples job submission logic.
The module now accepts a container of Sample objects on which to
iterate.  The execute method branches depending on whether data is being
retrieved from the Sample container or from a CSV file.  The code that
constructs the msub command is decoupled from execute.  There is also a
debug_flag class member that will not submit jobs via the Scheduler if
turned on.