Added TCGA logic, broke processing out into separate methods, perldoc.
Broke out the processing into separate methods, one for "default"
projects and the other to deal with TCGA, which requires some different
calculations and output formatting.  Also added project_name method and
class member to allow the class to branch accordingly.  Updated the
perldoc to be interspersed in the source code - more according to
convention, more likely to update, closer to what it describes, etc.

Fields:
	[+] project_name

Methods:
	[+] __base_output__
	[+] __tcga_output__
	[+] project_name