Adds frequency calculations to the joined frequency file.
This iterates on the HAPMAP files, counting the major homogenous
alleles, minor homogenous alleles, and heterogenous alleles for each RS
ID.  No-calls are ignored.  Then, matching against the RS IDs in the
joined frequency file, it combines that matched line with the frequency
calculations and produces the final frequency file.