Changed processing logic.
Figure out how the probelist allele is translated (if at all) in the
HAPMAP file, then do a corresponding translation to count the genotype
calls.

Read the probelist into a hash, then read the hapmap files, then iterate
on the hash to print the results.

Did away with the headers array because not all of the lines in the
HAPMAP files go by that.  Do a regex on each split column to get the
genotype calls instead.