Using Sample container for processing, block until Moab jobs complete.
Removed the input CSV file, everything is now passed in as the Sample
container, which is a subset of the Sample container processed on by
EGenoSolid which lacked CSFASTA files.  Also mimicked the dependency
list processing of the Framework scripts to wait to return until the
Moab submissions have completed.