Now passes correctly data to Bam2csfasta.
Sample objects are passed by reference to the container holding Sample
objects that need to be processed by Bam2csfasta.  This reference is
used later to update the underlying data structure, which caused
Bam2csfasta to throw an error.  Used Clone to do a deep copy of the
Sample reference to put into the Bam2csfasta container.